test and use UpdateFreezePeriod instead of SetFreePeriod
